pyproject.toml
[project]
name = "medquant"
version = "0.1.0"
description = "Benchmarking GGUF quantization effects on medical LLM QA with QLoRA fine-tuning"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "peft>=0.10.0",
    "trl>=0.8.6",
    "accelerate>=0.28.0",
    "bitsandbytes>=0.43.0",
    # GGUF inference
    "llama-cpp-python>=0.2.70",
    # Serving
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    # Utilities
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "scipy>=1.13.0",
    "huggingface_hub>=0.22.0",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
# Mac-only: install separately with: pip install mlx-lm
# mac = ["mlx-lm>=0.12.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
